ENH: Add CheckWindowsClassifier testing utility.
